requirements.txt
Flask~=2.2.2
python-dotenv~=0.21.1
pymongo~=4.3.3
bcrypt~=4.0.1
PyJWT~=2.6.0